requirements.txt
pyTelegramBotAPI>=3.5.1

pyYAML>=3.12
